+++ src/main.rs
@@ -1,4 +1,5 @@
 use rusqlite::{Connection, NO_PARAMS};
+use std::io::{self, Write};
 
 fn main() {
     let matches = cli::app().get_matches();
@@ -32,6 +33,30 @@ fn main() {
             column_params = column_params.join(", "),
         );
         conn.execute(&script, column_data).unwrap();
+    } else if let Some(matches) = matches.subcommand_matches("sub") {
+        let table_name = matches.value_of("list name").unwrap();
+
+        let id: String;
+        let (filter_names, filter_values) = if let Some(row_id) = matches.value_of("row ID") {
+            (vec!["id"], vec![row_id])
+        } else if let Some(filters) = matches.values_of("filters") {
+            column_partitioner(filters)
+        } else {
+            id = {
+                let mut stmt = conn.prepare(&query_builder::List::new(table_name, None).to_string()).unwrap();
+                let pretty_table = printer::prettytable(&mut stmt.query(NO_PARAMS).unwrap());
+                println!("{}", pretty_table);
+                print!("Enter the ID of the row to remove> ");
+                io::stdout().flush().unwrap();
+                let mut input = String::new();
+                io::stdin().read_line(&mut input).unwrap();
+                input
+            };
+            (vec!["id"], vec![id.as_str()])
+        };
+
+        let script = query_builder::Sub::new(table_name, filter_names).to_string();
+        conn.execute(&script, filter_values).unwrap();
     } else if let Some(matches) = matches.subcommand_matches("list") {
         let table_name = matches.value_of("list name").unwrap();
         let mut stmt;